Router added for browser navigation
