+++ package.json
@@ -10,7 +10,8 @@
         "lodash": "^4.17.21",
         "q": "^1.5.1",
         "react": "^17.0.2",
-        "react-dom": "^17.0.2"
+        "react-dom": "^17.0.2",
+        "react-router-dom": "^5.3.0"
     },
     "scripts": {
         "start": "react-scripts start",
